requirements.txt
django==4.0.5
psycopg2-binary==2.9.1
mailerlite_api_python==0.10.0
Pillow==9.1.0
celery==5.2.5
django_allauth==0.51.0